% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP.RST.R
\name{DP.RST}
\alias{DP.RST}
\title{DP-RST: Dirichlet Process Mixture of Random Spanning Trees}
\usage{
DP.RST(
  Y,
  graph0,
  init_val,
  hyperpar,
  MCMC,
  BURNIN,
  THIN,
  PT = TRUE,
  seed = 1234
)
}
\arguments{
\item{Y}{Standardized matrix of Principal Components.}

\item{graph0}{Spatial graph. Should be an \code{igraph} object containing \code{length(Y)} vertices.}

\item{init_val}{Named list of initial values containing:
\describe{
\item{\code{trees}}{List of initial spanning trees (M \code{igraph} objects).}
\item{\code{cluster}}{Initial cluster membership matrix (\code{length(Y) * M}).}
\item{\code{mu}}{List of initial cluster means.}
\item{\code{teams}}{Initial refined partition matrix (\code{length(Y) * M}).}
\item{\code{mu_teams}}{List of initial team means.}
\item{\code{sigmasq_y}}{Initial noise variance.}
}}

\item{hyperpar}{Named list of hyperparameters containing:
\describe{
\item{\code{M}}{Number of temperatures for Paralell Tempering. M=1 for one chain.}
\item{\code{sigmasq_mu}}{Variance of Gaussian prior for \code{mu}.}
\item{\code{lambda_s}}{Scale parameter of Inverse-Wishart prior.}
\item{\code{nu}}{Degrees of freedom for Inverse-Wishart prior.}
\item{\code{k_max}}{Maximum number of clusters in spatial partition.}
\item{\code{j_max}}{Maximum number of clusters in refined partition.}
\item{\code{temp}}{A vector of temperatures of size M for Parallel Tempering.}
}}

\item{MCMC}{Number of MCMC iterations.}

\item{BURNIN}{Number of burn-in iterations.}

\item{THIN}{Thinning interval (retains samples every \code{THIN} iterations).}

\item{PT}{Logical; whether to use Parallel Tempering.}

\item{seed}{Random seed.}
}
\value{
A list of MCMC samples:
\describe{
\item{\code{cluster_out}}{List of posterior cluster assignments.}
\item{\code{teams_out}}{List of refined cluster assignments.}
\item{\code{tree_out}}{List of spanning trees for each partition.}
\item{\code{k_out}}{Vector with the number of clusters in spatial partition.}
\item{\code{j_out}}{Vector with the number of clusters in refined partition.}
\item{\code{mu_out}}{List of posterior samples of \code{mu}.}
\item{\code{mu_teams_out}}{List of posterior samples of means for refined partition.}
\item{\code{sigmasq_y_out}}{List of posterior samples of noise variance.}
\item{\code{marginal_likelihood_out}}{Vector of marginal likelihood values.}
}
}
\description{
Implements the DP-RST algorithm using MCMC sampling.
}
