% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_boundary.R
\name{geom_boundary}
\alias{geom_boundary}
\title{Create a ggplot2 segment layer for a complex boundary}
\usage{
geom_boundary(bnd, ...)
}
\arguments{
\item{bnd}{A list with elements \code{x} and \code{y}, representing the boundary coordinates.}

\item{...}{Additional arguments passed to \code{ggplot2::geom_segment()}.}
}
\value{
A \code{ggplot2} \code{geom_segment} layer representing the boundary.
}
\description{
Create a ggplot2 segment layer for a complex boundary
}
