% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_plot.R
\name{groups_plot}
\alias{groups_plot}
\title{Plot Spatial Clusters}
\usage{
groups_plot(coords, group_assign, bound, point_size = 2)
}
\arguments{
\item{coords}{A matrix or dataframe with two columns representing X and Y coordinates.}

\item{group_assign}{A vector of group assignments corresponding to \code{coords}.}

\item{bound}{A spatial boundary object to overlay on the plot.}

\item{point_size}{The size of the points in the plot (default: 2).}
}
\value{
A ggplot object showing the spatial clusters.
}
\description{
Generates a scatter plot of spatial clusters, displaying the assigned groups
within the spatial boundary.
}
