% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraph.R
\name{plotGraph}
\alias{plotGraph}
\title{Plot a Graph Using ggplot2}
\usage{
plotGraph(coords, graph, title = NULL)
}
\arguments{
\item{coords}{A matrix or data frame with two columns representing node coordinates (x, y).}

\item{graph}{An \code{igraph} object representing the graph structure.}

\item{title}{A character string specifying the plot title (optional).}
}
\value{
A \code{ggplot2} object visualizing the graph edges.
}
\description{
This function visualizes a graph by plotting its edges based on provided coordinates.
}
\examples{
# Generate a mesh
coords <- matrix(runif(20), ncol = 2)
bnd <- list(x = c(0,1,1,0,0), y = c(0,0,1,1,0))
mesh <- gen2dMesh(coords, bnd)

# Create a constrained Delaunay triangulation graph
graph <- constrainedDentri(n = mesh$n_int, mesh = mesh, threshold = 0.5)

# Plot the graph
plotGraph(coords, graph, title = "Graph from Mesh")
# Plot the graph with the boundary
spatial_graph <- plotGraph(coords, graph, title = "Graph with Boundary from Mesh") +
                  geom_boundary(bnd)
spatial_graph
}
