% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_plot.R
\name{teams_plot}
\alias{teams_plot}
\title{Plot Refined Partition (Teams)}
\usage{
teams_plot(coords, team_assign, group_assign, bound, point_size = 2)
}
\arguments{
\item{coords}{A matrix or dataframe with two columns representing X and Y coordinates.}

\item{team_assign}{A vector of team assignments corresponding to \code{coords}.}

\item{group_assign}{A vector of group assignments corresponding to \code{coords}.}

\item{bound}{A spatial boundary object to overlay on the plot.}

\item{point_size}{The size of the points in the plot (default: 2).}
}
\value{
A ggplot object showing the refined partition (teams).
}
\description{
Generates a scatter plot of the refined partition, displaying assigned teams
within the spatial boundary.
}
